requirements.txt
pandas
numpy
mysql-connector-python
pymysql
psycopg2-binary
python-dotenv
schedule